backend/FarmAgent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
sqlmodel==0.0.21
apscheduler==3.10.4
firebase-admin==6.5.0